Refactoring Spine Runtime to meet demand

+++ spine-cpp/spine-cpp/include/spine/Extension.h
@@ -1,30 +1,31 @@
 /******************************************************************************
- * Spine Runtimes License Agreement
- * Last updated May 1, 2019. Replaces all prior versions.
+ * Spine Runtimes Software License v2.5
  *
- * Copyright (c) 2013-2019, Esoteric Software LLC
+ * Copyright (c) 2013-2016, Esoteric Software
+ * All rights reserved.
  *
- * Integration of the Spine Runtimes into software or otherwise creating
- * derivative works of the Spine Runtimes is permitted under the terms and
- * conditions of Section 2 of the Spine Editor License Agreement:
- * http://esotericsoftware.com/spine-editor-license
+ * You are granted a perpetual, non-exclusive, non-sublicensable, and
+ * non-transferable license to use, install, execute, and perform the Spine
+ * Runtimes software and derivative works solely for personal or internal
+ * use. Without the written permission of Esoteric Software (see Section 2 of
+ * the Spine Software License Agreement), you may not (a) modify, translate,
+ * adapt, or develop new applications using the Spine Runtimes or otherwise
+ * create derivative works or improvements of the Spine Runtimes or (b) remove,
+ * delete, alter, or obscure any trademarks or any copyright, trademark, patent,
+ * or other intellectual property or proprietary rights notices on or in the
+ * Software, including any copy thereof. Redistributions in binary or source
+ * form must include this license and terms.
  *
- * Otherwise, it is permitted to integrate the Spine Runtimes into software
- * or otherwise create derivative works of the Spine Runtimes (collectively,
- * "Products"), provided that each user of the Products must obtain their own
- * Spine Editor license and redistribution of the Products in any form must
- * include this license and copyright notice.
- *
- * THIS SOFTWARE IS PROVIDED BY ESOTERIC SOFTWARE LLC "AS IS" AND ANY EXPRESS
- * OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
- * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN
- * NO EVENT SHALL ESOTERIC SOFTWARE LLC BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES, BUSINESS
- * INTERRUPTION, OR LOSS OF USE, DATA, OR PROFITS) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
- * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
- * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ * THIS SOFTWARE IS PROVIDED BY ESOTERIC SOFTWARE "AS IS" AND ANY EXPRESS OR
+ * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
+ * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
+ * EVENT SHALL ESOTERIC SOFTWARE BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES, BUSINESS INTERRUPTION, OR LOSS OF
+ * USE, DATA, OR PROFITS) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
+ * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
  *****************************************************************************/
 
 #ifndef Spine_Extension_h
@@ -41,17 +42,17 @@ class String;
 class SP_API SpineExtension {
 public:
 	template<typename T>
-	static T *alloc(size_t num, const char *file, int line) {
+	static T *alloc(int32 num, const char *file, int line) {
 		return (T *) getInstance()->_alloc(sizeof(T) * num, file, line);
 	}
 
 	template<typename T>
-	static T *calloc(size_t num, const char *file, int line) {
+	static T *calloc(int32 num, const char *file, int line) {
 		return (T *) getInstance()->_calloc(sizeof(T) * num, file, line);
 	}
 
 	template<typename T>
-	static T *realloc(T *ptr, size_t num, const char *file, int line) {
+	static T *realloc(T *ptr, int32 num, const char *file, int line) {
 		return (T *) getInstance()->_realloc(ptr, sizeof(T) * num, file, line);
 	}
 
@@ -71,11 +72,11 @@ public:
 	virtual ~SpineExtension();
 
 	/// Implement this function to use your own memory allocator
-	virtual void *_alloc(size_t size, const char *file, int line) = 0;
+	virtual void *_alloc(int32 size, const char *file, int line) = 0;
 
-	virtual void *_calloc(size_t size, const char *file, int line) = 0;
+	virtual void *_calloc(int32 size, const char *file, int line) = 0;
 
-	virtual void *_realloc(void *ptr, size_t size, const char *file, int line) = 0;
+	virtual void *_realloc(void *ptr, int32 size, const char *file, int line) = 0;
 
 	/// If you provide a spineAllocFunc, you should also provide a spineFreeFunc
 	virtual void _free(void *mem, const char *file, int line) = 0;
@@ -96,11 +97,11 @@ public:
 	virtual ~DefaultSpineExtension();
 
 protected:
-	virtual void *_alloc(size_t size, const char *file, int line);
+	virtual void *_alloc(int32 size, const char *file, int line);
 
-	virtual void *_calloc(size_t size, const char *file, int line);
+	virtual void *_calloc(int32 size, const char *file, int line);
 
-	virtual void *_realloc(void *ptr, size_t size, const char *file, int line);
+	virtual void *_realloc(void *ptr, int32 size, const char *file, int line);
 
 	virtual void _free(void *mem, const char *file, int line);
